+++ NEWS.md
@@ -1,3 +1,19 @@
+# ggtangle 0.0.5
+
++ allows `color_edge = "category"` to color edges based on the category information in `cnetplot()` (2024-11-29, Fri)
++ update `cnetplot()` to address issues (2024-11-15, Fri)
+    - <https://github.com/YuLab-SMU/enrichplot/issues/293#issuecomment-2462502566>
++ update vignette to add a session of `cnetplot()` (2024-11-06, Wed)
++ 'node_label' can be a vector of selected items/genes to specify the items displayed in `cnetplot()` (2024-11-06, Wed)
+    - if `options(cnetplot_subset = TRUE)`, all the edges and nodes of the genes that are not wanted to labeled will be remove 
++ `geom_cnet_label()` allows detail adjustment of `cnetplot()` labels (2024-11-06, Wed)
+    - <https://github.com/YuLab-SMU/enrichplot/issues/194>
++ 'node_label' in `cnetplot()` supports new arguments (2024-11-05, Tue)
+    - 'exclusive' to label genes that is uniquely belong to categories
+    - 'share' to label genes that are shared between categories
+    - expresson, like '> 1' or '< 1', to label those genes with `foldChange > 1` or `foldChange < 1`
+    - <https://github.com/YuLab-SMU/enrichplot/issues/253>
+
 # ggtangle 0.0.4
 
 + compatible with more igraph layouts (2024-10-29, Tue)